backend/utilisateur/serializer.py
from rest_framework import serializers
from .models import Utilisateur


class UtilisateurSerializer(serializers.ModelSerializer):
    class Meta:
        model = Utilisateur
        fields = [
            'id',
            'username',
            'email',
            'first_name',
            'last_name',
            'photo_de_profile',
            'description',
            'age',
            'lien_cv',
            'telephone',
        ]